api/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 